back/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-braces==1.14.0
django-cors-headers==3.2.1
django-extensions==2.2.6
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-smtp-ssl==1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
rest-social-auth==3.0.0
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
